The arc command is updated.
The option `delta` is added, and `start` and `stop` are not obligatory.
But two of the three should be defined, if not an error will be given.
Fixed an error that the start angle could not be less than the stop
angle (when computing the samples, the number was negative). The manual
is updated and the `mode` option is added to the manual.