Support Content Intersections (#386)
- Adds intersection checking for `content` elements by setting their
`segments` to their frame (defaulting to a rectangle)
- Fixes a bug in `line` intersection checking, taking the furthest
intersection for one side and the nearest for the other

This changes `intersection` to _not_ emit anchors for intersections of
one element's drawables with each other (e.g. an element with 2
drawables or marks did emit self intersections prior to this change!).
See the grid intersection example (ref image), which should _not_ emit
intersections.