In tree, add "end" and "after-end" as values for parent-position (#444)
This solves the issue I created last week (#429) and adds another layout
choice where the parent is offset relative to the children.

Both values seem more useful for "right" and "left" direction, and
actually, the "after-end" option seems more useful to me, as it allows
to render a more hierarchical layout without overlapping, like this
question asked on Discord:
https://discord.com/channels/1054443721975922748/1193008120688037988

I did not change the existing behavior, but it was very surprising to me
to see that when choosing "right" or "left" as direction, the nodes at
filled from bottom to top, and it is necessary to choose "end" to have
the parent on top. Is it that way for a reason? I would prefer a
"before-beginning" rather than "after-end" to indicate that I want my
parent above the children, it seems more intuitive to me.