Fix: Prevent divide by zero error if a segment has zero length (#495)
I found this bug by attempting to use a path anchor on a group with a
vertical line:
```typc
group(name: "g", {
  line((0,0), (0,1))
})
circle("g.start")
```
The group's width becomes zero so some of the points on its bounds are
equal. This causes their length to be zero, which causes a divide zero
error in `path-util._point-on-line-segment`.

I have fixed this by checking if length is a zero integer and setting it
to 1 if so. I'm not sure what other issues this may cause...