shapes: Add bbox primitive (rect as aabb)
Drawing a border around several elements is a common task when creating
scientific drawings to indicate logical connection. The bbox shape is a
wrapper for the existing rect primitive that calculates an axis aligned
bounding box for a given collection of coordinates or elements.
If a shape is given, all of its anchors are taken into account for
aabb calculation.

The padding attribute can be used to specify inner spacing of the
bounding box. All other styling attributes are forwarded to the rect.