docs: Continue fixing `tidy` and add consistency
This commit continues the rework of the documentation.

This time, all changes have been tested by building the actual site and
they all seem to be working fairly well.

Further work will continue on the rest of the major sections of the
documentation site, as well as on whatever points are yet to be answered
by official mantainers.