perf: Finish fn signature and parameter parser
Following the plan in the PR this branch is part of, the function
signature parser rework is done. The conclusion is that the signature
gets parsed now all in a single pass instead of having to separately
consider the function and parameter span, and then parse the parameter
span.

The gains in efficency on the prior commit are still kept, so no
operations are performed on the whole array; only those elements of the
array with all the source lines corresponding to the function are
parsed.

There is, though, a small performance loss in the fact that a custom
runtime regex is built to more accurately have the whitespace
indentation of some parameters' default values be represented in the
final output. Because the project doesn't use an autoformatter, if some
lines contain an indentation that is beyond a single multiple of 2
(according to the .editorconfig file) then the parser will correctly
recognize that the contents of the named argument, if the named argument
is not a string or Typst content type, should be deindented by however
as much whitespace was detected at the start of the parameter name.

Now the parameter parser also correctly implements function default
arguments.